requirements.txt
openpyxl>=3.0.0
pillow>=8.0.0